+++ src/Facades/Auth.php
@@ -7,8 +7,6 @@ use Illuminate\Support\Facades\Facade;
 /**
  * Class Saml2Auth
  *
- * @method static \Slides\Saml2\Models\Tenant|null getTenant()
- *
  * @package Slides\Saml2\Facades
  */
 class Auth extends Facade