+++ src/Saml2User.php
@@ -3,7 +3,6 @@
 namespace Slides\Saml2;
 
 use OneLogin\Saml2\Auth as OneLoginAuth;
-use Slides\Saml2\Models\Tenant;
 
 /**
  * Class Saml2User
@@ -20,22 +19,13 @@ class Saml2User
     protected $auth;
 
     /**
-     * The tenant user belongs to.
-     *
-     * @var Tenant
-     */
-    protected $tenant;
-
-    /**
      * Saml2User constructor.
      *
      * @param OneLoginAuth $auth
-     * @param Tenant $tenant
      */
-    public function __construct(OneLoginAuth $auth, Tenant $tenant)
+    public function __construct(OneLoginAuth $auth)
     {
         $this->auth = $auth;
-        $this->tenant = $tenant;
     }
 
     /**
@@ -163,25 +153,4 @@ class Saml2User
         return $this->auth->getNameId();
     }
 
-    /**
-     * Set a tenant
-     *
-     * @param Tenant $tenant
-     *
-     * @return void
-     */
-    public function setTenant(Tenant $tenant)
-    {
-        $this->tenant = $tenant;
-    }
-
-    /**
-     * Get a resolved tenant.
-     *
-     * @return Tenant|null
-     */
-    public function getTenant()
-    {
-        return $this->tenant;
-    }
 }