+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+## [UNRELEASED]
+
+### Changed
+- IdPs configuration is now stored in the main configuration file
+- IdPs are identified by a key instead of UUID 
+- Name ID format configurable
+
+### Removed
+- Tenant model and repository
+- CLI commands
+
+
 ## [2.4.0] - 2024-04-13
 
 ### Added